Refactor inference module into submodules and add multilabel support
