+++ src/lisbet/inference/__init__.py
@@ -1,0 +1,7 @@
+from lisbet.inference.annotation import annotate_behavior
+from lisbet.inference.embedding import compute_embeddings
+
+__all__ = ["annotate_behavior", "compute_embeddings"]
+__doc__ = """
+Inference utilities for LISBET, including behavior annotation and embedding extraction.
+"""
